Bug 1734178: allow user to delete own account
added settings

- institution level: reviewselfdeletion
    0 if the institution does not require approval
      from an admin to delete an account
    1 if the institution requires an admin to approve
      account deletion requests from users
    if not set, it takes the value from the site's
    default

- site level: defaultreviewselfdeletion
  (Site options->User Settings -> Review account before self-deletion)
    1 if the site's default is requiring approval
    null otherwise

Account deletion by a user

when a user accesses to the account settings, a
'Delete account' button is displayed.

This will:
- If the user belongs to an institution that requires
  approval (or does not have the settings but the site
  requires approval by default)
    then a notification will be sent to the admins
    of the institutions that require approval that
    the user belongs to
- if the user belongs to institutions and none of them
  require approval (or does not have the setting
  but the site does not require approval by default)
    then the account is deleted
- if the user does not belong to any institution
    then the action will depend on the setting of
    the 'mahara' institution or sites default if
    'mahara' doesn't have the setting

Approval by institution admins

An institution admin can see the pending deletion
requests in Admin menu-> Institution -> Pending deletions
After approving/denying a request, the user
that requested the account deletion will receive
a notification

behatnotneeded

+++ htdocs/account/delete.php
@@ -13,34 +13,77 @@ define('INTERNAL', 1);
 define('MENUITEM', 'settings/preferences');
 
 require(dirname(dirname(__FILE__)) . '/init.php');
-define('TITLE', get_string('deleteaccount', 'account'));
+define('TITLE', get_string('deleteaccount', 'account', display_name($USER, null, false, false, true)));
 
 if (!$USER->can_delete_self()) {
     throw new AccessDeniedException(get_string('accessdenied', 'error'));
 }
 
-$deleteform = pieform(array(
+$deleteform = array(
     'name' => 'account_delete',
     'plugintype' => 'core',
     'pluginname' => 'account',
-    'elements'   => array(
+);
+
+$userid = $USER->get('id');
+$user = new User;
+$user->find_by_id($userid);
+$requiresapproval = $user->requires_delete_approval();
+
+if ($requiresapproval) {
+    $elements = array(
+        'reason' => array(
+            'type'  => 'textarea',
+            'title' => get_string('reason'),
+            'cols'  => 50,
+            'rows'  => 4,
+            'rules' => array('required' => true),
+        ),
         'submit' => array(
             'class' => 'btn-default',
-            'type' => 'submit',
-            'value' => get_string('deleteaccount', 'mahara', display_username($USER), full_name($USER)),
+            'type' => 'submitcancel',
+            'value' => array(get_string('senddeletenotification', 'mahara'), get_string('back')),
+            'goto' => get_config('wwwroot'). 'account/index.php',
         ),
-    ),
-));
+    );
+}
+else {
+  $elements = array(
+      'submit' => array(
+          'class' => 'btn-default',
+          'type' => 'submitcancel',
+          'value' => array(get_string('deleteaccount1', 'mahara'), get_string('back')),
+          'goto' => get_config('wwwroot'). 'account/index.php',
+      ),
+  );
+}
+
+$deleteform['elements'] = $elements;
+
+$deleteform = pieform($deleteform);
 
 function account_delete_submit(Pieform $form, $values) {
-    global $SESSION, $USER;
+    global $SESSION, $USER, $user;
     $userid = $USER->get('id');
-    $USER->logout();
-    delete_user($userid);
-    $SESSION->add_ok_msg(get_string('accountdeleted', 'account'));
-    redirect('/index.php');
+
+    // check if user needs approval to delete its account
+    if (!$user->requires_delete_approval()) {
+        $USER->logout();
+        delete_user($userid);
+        $SESSION->add_ok_msg(get_string('accountdeleted', 'account'));
+    }
+    else {
+        $admins = $user->get_approval_admins();
+        set_account_pending_deletion($userid, strip_tags(clean_html($values['reason'])));
+        $user->notify_admins_pending_deletion($admins, $values['reason']);
+        $SESSION->add_ok_msg(get_string('pendingdeletionemailsent', 'account'));
+    }
+    redirect('/account/index.php');
 }
 
 $smarty = smarty();
+$smarty->assign('requiresapproval', $requiresapproval);
 $smarty->assign('delete_form', $deleteform);
+$smarty->assign('fullname', full_name($USER));
+$smarty->assign('displayusername', display_username($USER));
 $smarty->display('account/delete.tpl');
